+++ contracts/modules/Checkpoint/ERC20DividendCheckpoint.sol
@@ -124,9 +124,6 @@ contract ERC20DividendCheckpoint is DividendCheckpoint {
         uint256 dividendIndex = dividends.length;
         uint256 currentSupply = ISecurityToken(securityToken).totalSupplyAt(_checkpointId);
         uint256 excludedSupply = 0;
-        for (uint256 i = 0; i < _excluded.length; i++) {
-            excludedSupply = excludedSupply.add(ISecurityToken(securityToken).balanceOfAt(_excluded[i], _checkpointId));
-        }
         dividends.push(
           Dividend(
             _checkpointId,
@@ -135,7 +132,7 @@ contract ERC20DividendCheckpoint is DividendCheckpoint {
             _expiry,
             _amount,
             0,
-            currentSupply.sub(excludedSupply),
+            0,
             false,
             0,
             0,
@@ -145,12 +142,14 @@ contract ERC20DividendCheckpoint is DividendCheckpoint {
 
         for (uint256 j = 0; j < _excluded.length; j++) {
             require (_excluded[j] != address(0), "Invalid address");
-            for (i = j + 1; i < _excluded.length; i++) {
+            for (uint256 i = j + 1; i < _excluded.length; i++) {
                 require (_excluded[j] != _excluded[i], "Duplicate exclude address");
+                excludedSupply = excludedSupply.add(ISecurityToken(securityToken).balanceOfAt(_excluded[i], _checkpointId));
+                dividends[dividends.length - 1].dividendExcluded[_excluded[i]] = true;
             }
             dividends[dividends.length - 1].dividendExcluded[_excluded[j]] = true;
         }
-
+        dividends[dividends.length - 1].totalSupply = currentSupply.sub(excludedSupply);
         dividendTokens[dividendIndex] = _token;
         _emitERC20DividendDepositedEvent(_checkpointId, _maturity, _expiry, _token, _amount, currentSupply, dividendIndex, _name);
     }