+++ contracts/modules/Checkpoint/EtherDividendCheckpoint.sol
@@ -106,9 +106,6 @@ contract EtherDividendCheckpoint is DividendCheckpoint {
         uint256 dividendIndex = dividends.length;
         uint256 currentSupply = ISecurityToken(securityToken).totalSupplyAt(_checkpointId);
         uint256 excludedSupply = 0;
-        for (uint256 i = 0; i < _excluded.length; i++) {
-            excludedSupply = excludedSupply.add(ISecurityToken(securityToken).balanceOfAt(_excluded[i], _checkpointId));
-        }
         dividends.push(
           Dividend(
             _checkpointId,
@@ -117,7 +114,7 @@ contract EtherDividendCheckpoint is DividendCheckpoint {
             _expiry,
             msg.value,
             0,
-            currentSupply.sub(excludedSupply),
+            0,
             false,
             0,
             0,
@@ -127,12 +124,14 @@ contract EtherDividendCheckpoint is DividendCheckpoint {
 
         for (uint256 j = 0; j < _excluded.length; j++) {
             require (_excluded[j] != address(0), "Invalid address");
-            for (i = j + 1; i < _excluded.length; i++) {
+            for (uint256 i = j + 1; i < _excluded.length; i++) {
                 require (_excluded[j] != _excluded[i], "Duplicate exclude address");
+                excludedSupply = excludedSupply.add(ISecurityToken(securityToken).balanceOfAt(_excluded[i], _checkpointId));            
+                dividends[dividends.length - 1].dividendExcluded[_excluded[i]] = true;
             }
             dividends[dividends.length - 1].dividendExcluded[_excluded[j]] = true;
         }
-
+        dividends[dividends.length -1].totalSupply = currentSupply.sub(excludedSupply);
         emit EtherDividendDeposited(msg.sender, _checkpointId, now, _maturity, _expiry, msg.value, currentSupply, dividendIndex, _name);
     }